Address PR feedback
* refactor into a profiling.rs module
* remove lazy_Static and mut_static
* replace mutex with AtomicBool
* remove profiling feature from test
